Code restructure to accomodate for firebase implementation
