+++ lib/Screens/main.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:rfg_app/homeScreen.dart';
-import 'package:rfg_app/settings.dart';
-import 'package:rfg_app/listScreen.dart';
+import 'package:rfg_app/Screens/homeScreen.dart';
+import 'package:rfg_app/Screens/settings.dart';
+import 'package:rfg_app/Screens/listScreen.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'firebase_options.dart';
