+++ pubspec.yaml
@@ -41,6 +41,8 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
   firebase_core: ^2.7.1
+  firebase_auth: ^4.2.2
+  cloud_firestore: ^4.4.4
 
 dev_dependencies:
   flutter_test:
